LoongArch: create trampoline for short function call
when a function call the other file in the same .o,it always
use short jump like b/bl,but when the other function is not
included in the output diff object,the short jump is not enough.
create a .text.kpatch_trampoline section to indirectly jump to
the target remote function.

Signed-off-by: Hongchen Zhang <zhanghongchen@loongson.cn>